+++ source/common/upstream/cluster_factory_impl.cc
@@ -58,11 +58,13 @@ std::pair<ClusterSharedPtr, ThreadAwareLoadBalancerPtr> ClusterFactoryImplBase::
     cluster_type = cluster.cluster_type().name();
   }
 
-  if (cluster.common_lb_config().use_hostname() &&
+  if (cluster.common_lb_config().has_consistent_hashing_lb_config() &&
+      cluster.common_lb_config().consistent_hashing_lb_config().use_hostname_for_hashing() &&
       !(cluster.type() == envoy::config::cluster::v3::Cluster::STRICT_DNS ||
         cluster.type() == envoy::config::cluster::v3::Cluster::LOGICAL_DNS)) {
-    throw EnvoyException(
-        fmt::format("Cannot use hostname loadbalancing for cluster of type: '{}'", cluster_type));
+    throw EnvoyException(fmt::format(
+        "Cannot use hostname for consistent hashing loadbalancing for cluster of type: '{}'",
+        cluster_type));
   }
   ClusterFactory* factory = Registry::FactoryRegistry<ClusterFactory>::getFactory(cluster_type);