+++ source/common/upstream/maglev_lb.h
@@ -34,7 +34,7 @@ class MaglevTable : public ThreadAwareLoadBalancerBase::HashingLoadBalancer,
                     Logger::Loggable<Logger::Id::upstream> {
 public:
   MaglevTable(const NormalizedHostWeightVector& normalized_host_weights,
-              double max_normalized_weight, uint64_t table_size, bool use_hostname,
+              double max_normalized_weight, uint64_t table_size, bool use_hostname_for_hashing,
               MaglevLoadBalancerStats& stats);
 
   // ThreadAwareLoadBalancerBase::HashingLoadBalancer
@@ -82,7 +82,7 @@ private:
   createLoadBalancer(const NormalizedHostWeightVector& normalized_host_weights,
                      double /* min_normalized_weight */, double max_normalized_weight) override {
     return std::make_shared<MaglevTable>(normalized_host_weights, max_normalized_weight,
-                                         table_size_, use_hostname_, stats_);
+                                         table_size_, use_hostname_for_hashing_, stats_);
   }
 
   static MaglevLoadBalancerStats generateStats(Stats::Scope& scope);
@@ -90,7 +90,7 @@ private:
   Stats::ScopePtr scope_;
   MaglevLoadBalancerStats stats_;
   const uint64_t table_size_;
-  const bool use_hostname_;
+  const bool use_hostname_for_hashing_;
 };
 
 } // namespace Upstream